requirements.txt
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
cryptography==1.7.2
Django==2.0.2
falcon==1.4.1
hug==2.4.0
idna==2.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.1
python-mimeparse==1.6.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
